+++ courtlistener/__init__.py
@@ -1,3 +1,4 @@
+from courtlistener.citation_lookup import CitationLookup
 from courtlistener.client import CourtListener
 
-__all__ = ["CourtListener"]
+__all__ = ["CitationLookup", "CourtListener"]